+++ src/imio/smartweb/core/browser/static/src/edit.js
@@ -5,6 +5,33 @@ jQuery(document).ready(function ($) {
     $(".hide-in-preview, #section-byline, #global_statusmessage").toggle("fast");
     e.preventDefault();
   });
+
+  // Uncheck checked icon
+  $("#formfield-form-widgets-svg_icon input").click(function(e){
+    var $elm_clicked = $(this);
+    if ($(this).attr('checked')) {
+      $(this).prop("checked", false );
+      $(this).removeAttr("checked");
+      $(this).css("box-shadow", "none");
+      $(this).css("border-color", "#DEE2ED");
+    }
+    else {
+      $("#formfield-form-widgets-svg_icon input").each(function(index, elm) {
+        if ($elm_clicked[0] === elm) {
+          $(this).prop("checked", true );
+          $(this).attr("checked", "checked");
+          $(this).css("border-color", "#007a99");
+          $(this).css("box-shadow", "0 0 0 0.25rem rgb(0 122 153 / 25%)");
+        }
+        else {
+          $(this).prop("checked", false );
+          $(this).removeAttr("checked");
+          $(this).css("box-shadow", "none");
+          $(this).css("border-color", "#DEE2ED");
+        }
+      })
+    }
+  });
 });
 
 jQuery(window).on("load", function(e) {