+++ CHANGES.rst
@@ -113,6 +113,8 @@ Changelog
 
 - Adaptation of react to show or hide the map
   [thomlamb]
+- WEB-3992 : Uncheck icon when clincking on checked icon (in edit form of imio.smartweb.BlockLink)
+  [boulch]
 
 - WEB-3999 : Keep order of contacts in its view through manualy sorted related_contacts in edit form
   [boulch]